refactor(heureka): move ServiceType to a centeral place
