+++ apps/heureka/src/components/Services/ServicesList/ServiceListItem/ServiceListItem.tsx
@@ -4,9 +4,10 @@
  */
 
 import React from "react"
-import { DataGridRow, DataGridCell, Pill, Badge, Stack, Button } from "@cloudoperators/juno-ui-components"
-import { ServiceType } from "../.."
-import { SeverityCount } from "../../../common/IssuesCountBadges/IssuesCount"
+import { DataGridRow, DataGridCell, Pill, Stack, Button } from "@cloudoperators/juno-ui-components"
+import { SeverityCount } from "../../../common/SeverityCount"
+import { ServiceType } from "../../../types"
+
 type ServiceDetailsLabel = {
   [key: string]: string
 }
